038.lambda-layer-test/scripts/requirements.txt
pydantic==2.12.4
psycopg2-binary==2.9.11